+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/jeertmans/languagetool-rust/compare/v3.0.1...HEAD)
 
+### Fixed
+
+- Bring back --more-context flag
+
 ## [3.0.1](https://github.com/jeertmans/languagetool-rust/compare/v3.0.0...3.0.1) 2025-12-31
 
 ### Changed